pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel", "torch"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint.isort]
lines-between-types = 1
lines-after-imports = 2

[project]
name = "vipe"
description = "NVIDIA Video Pose Engine"
authors = [
    {name = "The ViPE Authors", email = "jiahuih@nvidia.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "numpy",
    "einops",
    "hydra-core",
    "omegaconf",
    "pandas",
    "scikit-learn",
    "torch-scatter",
    "tqdm",
    "rerun-sdk",
    "imageio[ffmpeg]",
    "opencv-python",
    "python-pycg",
    "timm",
    "transformers",
    "gdown",
    "OpenEXR",
    "kornia",
    "viser",
    "click",
]

[project.scripts]
vipe = "vipe.cli.main:main"
